# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopmee API Documentation",
        "description": "API endpoints for Shopmee e-commerce application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Shopmee API Server"
        }
    ],
    "paths": {
...
        "/shops/register": {
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Đăng ký mở cửa hàng (Shop)",
                "description": "Đăng ký thông tin cửa hàng cho tài khoản hiện tại.",
                "operationId": "registerShop",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Shop Quan Ao Dep"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Chuyên cung cấp quần áo thời trang"
                                    },
                                    "logo_url": {
                                        "type": "string",
                                        "example": "https://example.com/logo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký mở cửa hàng thành công!"
                                        },
                                        "data": {
                                            "properties": {
                                                "shop_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ hoặc tài khoản đã có Shop"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}